requirements.txt
langchain==0.2.5
langchain-community==0.2.5
chromadb==0.5.3
pypdf==4.2.0
sentence-transformers==3.0.1
anthropic==0.86.0
fastapi==0.111.0
uvicorn==0.30.1
python-dotenv==1.0.1
ragas==0.1.9
datasets==2.19.1
rank-bm25==0.2.2
openai